+++ Content.Server/_Citadel/Overmaps/Systems/OvermapSystem.API.cs
@@ -1,3 +1,4 @@
+using Content.Shared._Citadel.Overmaps;
 using Content.Shared._Citadel.Overmaps.Systems;
 
 namespace Content.Server._Citadel.Overmaps.Systems
