+++ Content.Shared/_Citadel/Overmaps/Systems/SharedOvermapSystem.API.cs
@@ -3,7 +3,7 @@ using Content.Shared._Citadel.Overmaps.Components;
 
 namespace Content.Shared._Citadel.Overmaps.Systems
 {
-    public sealed partial class SharedOvermapSystem : EntitySystem
+    public partial class SharedOvermapSystem : EntitySystem
     {
         public bool TryGetOvermap([NotNullWhen(true)] OvermapId? overmapId,
             [NotNullWhen(true)] out EntityUid? uid)
